Discord Bot base
the base of any discord bot, when ran it prints "Ready!" in the console and does nothing else.  The prefix is set to "!" but can be changed later.  Only the file "index.js" should be edited and to run this bot, type "node index.js" while in the directory with the index.js file.