links question & answers, adds and subtracts points
generates a 6 digit 0-9a-z code for each question and allows users to answer questions using that code.  Also adds points if an upvote was added or a downvote was removed and removes points if a downvote was added or an upvote was removed.

Uses functions (structs) instead of dictionaries for storing information